+++ qiskit/visualization/circuit/__init__.py
@@ -10,10 +10,6 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""
-Module for the primary interface to the circuit drawers.
-"""
+""" Init for circuit visualizations """
 
-# Temporary import from 0.22.0 to be deprecated in future
-# pylint: disable=unused-import
-from .circuit.circuit_visualization import circuit_drawer
+from .circuit_visualization import circuit_drawer
