+++ qiskit/visualization/circuit/styles/bw.json
