+++ qiskit/visualization/circuit/styles/default.json
