+++ qiskit/visualization/circuit/styles/iqx.json
