+++ qiskit/visualization/circuit/styles/textbook.json
