+++ releasenotes/notes/visualization-reorganisation-9e302239705c7842.yaml
@@ -1,0 +1,9 @@
+---
+upgrade:
+  - |
+    The visualization module :mod:`qiskit.visualization` has seen some internal
+    reorganisation.  This should not have affected the public interface, but if
+    you were accessing any internals of the circuit drawers, they may now be in
+    different places.  The only parts of the visualization module that are
+    considered public are the components that are documented in this online
+    documentation.
