+++ test/ipynb/mpl/graph/test_graph_matplotlib_drawer.py
@@ -18,11 +18,11 @@ import json
 import os
 from contextlib import contextmanager
 
-from qiskit.visualization.state_visualization import state_drawer
 from qiskit import BasicAer, execute
 from qiskit.test import QiskitTestCase
 from qiskit import QuantumCircuit
 from qiskit.utils import optionals
+from qiskit.visualization.state_visualization import state_drawer
 from qiskit.visualization.counts_visualization import plot_histogram
 from qiskit.visualization.gate_map import plot_gate_map, plot_coupling_map
 from qiskit.providers.fake_provider import (
