+++ test/python/visualization/test_circuit_text_drawer.py
@@ -25,8 +25,8 @@ from qiskit.quantum_info.operators import SuperOp
 from qiskit.quantum_info.random import random_unitary
 from qiskit.test import QiskitTestCase
 from qiskit.transpiler import Layout
-from qiskit.visualization import text as elements
-from qiskit.visualization.circuit_visualization import _text_circuit_drawer
+from qiskit.visualization.circuit import text as elements
+from qiskit.visualization.circuit.circuit_visualization import _text_circuit_drawer
 from qiskit.extensions import UnitaryGate, HamiltonianGate
 from qiskit.extensions.quantum_initializer import UCGate
 from qiskit.circuit.library import (
