+++ test/python/visualization/test_dag_drawer.py
@@ -19,9 +19,9 @@ import unittest
 from PIL import Image
 
 from qiskit.circuit import QuantumRegister, QuantumCircuit, Qubit, Clbit
-from qiskit.tools.visualization import dag_drawer
+from qiskit.visualization import dag_drawer
 from qiskit.exceptions import InvalidFileError
-from qiskit.visualization.exceptions import VisualizationError
+from qiskit.visualization import VisualizationError
 from qiskit.converters import circuit_to_dag
 from qiskit.utils import optionals as _optionals
 from .visualization import path_to_diagram_reference, QiskitVisualizationTestCase
@@ -41,7 +41,8 @@ class TestDagDrawer(QiskitVisualizationTestCase):
     @unittest.skipUnless(_optionals.HAS_GRAPHVIZ, "Graphviz not installed")
     def test_dag_drawer_invalid_style(self):
         """Test dag draw with invalid style."""
-        self.assertRaises(VisualizationError, dag_drawer, self.dag, style="multicolor")
+        with self.assertRaisesRegex(VisualizationError, "Invalid style multicolor"):
+            dag_drawer(self.dag, style="multicolor")
 
     @unittest.skipUnless(_optionals.HAS_GRAPHVIZ, "Graphviz not installed")
     def test_dag_drawer_checks_filename_correct_format(self):
