+++ test/python/visualization/test_utils.py
@@ -17,14 +17,13 @@ import numpy as np
 
 from qiskit import QuantumRegister, ClassicalRegister, QuantumCircuit
 from qiskit.circuit import Qubit, Clbit
-from qiskit.visualization import utils, array_to_latex
+from qiskit.visualization.circuit import _utils
+from qiskit.visualization import array_to_latex
 from qiskit.test import QiskitTestCase
 
 
 class TestVisualizationUtils(QiskitTestCase):
-    """Tests for visualizer utilities.
-    Since the utilities in qiskit/tools/visualization/_utils.py are used by several visualizers
-    the need to be check if the interface or their result changes."""
+    """Tests for circuit drawer utilities."""
 
     def setUp(self):
         super().setUp()
@@ -45,7 +44,7 @@ class TestVisualizationUtils(QiskitTestCase):
 
     def test_get_layered_instructions(self):
         """_get_layered_instructions without reverse_bits"""
-        (qregs, cregs, layered_ops) = utils._get_layered_instructions(self.circuit)
+        (qregs, cregs, layered_ops) = _utils._get_layered_instructions(self.circuit)
 
         exp = [
             [("cx", (self.qr2[0], self.qr2[1]), ()), ("cx", (self.qr1[0], self.qr1[1]), ())],
@@ -64,7 +63,7 @@ class TestVisualizationUtils(QiskitTestCase):
 
     def test_get_layered_instructions_reverse_bits(self):
         """_get_layered_instructions with reverse_bits=True"""
-        (qregs, cregs, layered_ops) = utils._get_layered_instructions(
+        (qregs, cregs, layered_ops) = _utils._get_layered_instructions(
             self.circuit, reverse_bits=True
         )
 
@@ -100,7 +99,7 @@ class TestVisualizationUtils(QiskitTestCase):
         circuit.cx(qr1[1], qr1[0])
         circuit.measure(qr1[1], cr1[1])
 
-        (qregs, cregs, layered_ops) = utils._get_layered_instructions(circuit, idle_wires=False)
+        (qregs, cregs, layered_ops) = _utils._get_layered_instructions(circuit, idle_wires=False)
 
         exp = [
             [("cx", (qr2[0], qr2[1]), ()), ("cx", (qr1[0], qr1[1]), ())],
@@ -133,7 +132,7 @@ class TestVisualizationUtils(QiskitTestCase):
         qc.h(2)
         qc.cx(0, 3)
 
-        (_, _, layered_ops) = utils._get_layered_instructions(qc, justify="left")
+        (_, _, layered_ops) = _utils._get_layered_instructions(qc, justify="left")
 
         l_exp = [
             [
@@ -163,7 +162,7 @@ class TestVisualizationUtils(QiskitTestCase):
         qc.h(2)
         qc.cx(0, 3)
 
-        (_, _, layered_ops) = utils._get_layered_instructions(qc, justify="right")
+        (_, _, layered_ops) = _utils._get_layered_instructions(qc, justify="right")
 
         r_exp = [
             [("cx", (Qubit(QuantumRegister(4, "q"), 0), Qubit(QuantumRegister(4, "q"), 3)), ())],
@@ -212,7 +211,7 @@ class TestVisualizationUtils(QiskitTestCase):
         """
         qc = QuantumCircuit.from_qasm_str(qasm)
 
-        (_, _, layered_ops) = utils._get_layered_instructions(qc, justify="left")
+        (_, _, layered_ops) = _utils._get_layered_instructions(qc, justify="left")
 
         l_exp = [
             [
@@ -279,7 +278,7 @@ class TestVisualizationUtils(QiskitTestCase):
         """
         qc = QuantumCircuit.from_qasm_str(qasm)
 
-        (_, _, layered_ops) = utils._get_layered_instructions(qc, justify="right")
+        (_, _, layered_ops) = _utils._get_layered_instructions(qc, justify="right")
 
         r_exp = [
             [
@@ -332,7 +331,7 @@ class TestVisualizationUtils(QiskitTestCase):
         qc_2.measure(0, 0)
         qc.append(qc_2, [1], [0])
 
-        (_, _, layered_ops) = utils._get_layered_instructions(qc)
+        (_, _, layered_ops) = _utils._get_layered_instructions(qc)
 
         expected = [
             [("h", (Qubit(QuantumRegister(2, "q"), 0),), ())],
@@ -358,37 +357,38 @@ class TestVisualizationUtils(QiskitTestCase):
 
     def test_generate_latex_label_nomathmode(self):
         """Test generate latex label default."""
-        self.assertEqual("abc", utils.generate_latex_label("abc"))
+        self.assertEqual("abc", _utils.generate_latex_label("abc"))
 
     def test_generate_latex_label_nomathmode_utf8char(self):
         """Test generate latex label utf8 characters."""
         self.assertEqual(
-            "{\\ensuremath{\\iiint}}X{\\ensuremath{\\forall}}Y", utils.generate_latex_label("∭X∀Y")
+            "{\\ensuremath{\\iiint}}X{\\ensuremath{\\forall}}Y",
+            _utils.generate_latex_label("∭X∀Y"),
         )
 
     def test_generate_latex_label_mathmode_utf8char(self):
         """Test generate latex label mathtext with utf8."""
         self.assertEqual(
             "abc_{\\ensuremath{\\iiint}}X{\\ensuremath{\\forall}}Y",
-            utils.generate_latex_label("$abc_$∭X∀Y"),
+            _utils.generate_latex_label("$abc_$∭X∀Y"),
         )
 
     def test_generate_latex_label_mathmode_underscore_outside(self):
         """Test generate latex label with underscore outside mathmode."""
         self.assertEqual(
             "abc_{\\ensuremath{\\iiint}}X{\\ensuremath{\\forall}}Y",
-            utils.generate_latex_label("$abc$_∭X∀Y"),
+            _utils.generate_latex_label("$abc$_∭X∀Y"),
         )
 
     def test_generate_latex_label_escaped_dollar_signs(self):
         """Test generate latex label with escaped dollarsign."""
-        self.assertEqual("${\\ensuremath{\\forall}}$", utils.generate_latex_label(r"\$∀\$"))
+        self.assertEqual("${\\ensuremath{\\forall}}$", _utils.generate_latex_label(r"\$∀\$"))
 
     def test_generate_latex_label_escaped_dollar_sign_in_mathmode(self):
         """Test generate latex label with escaped dollar sign in mathmode."""
         self.assertEqual(
             "a$bc_{\\ensuremath{\\iiint}}X{\\ensuremath{\\forall}}Y",
-            utils.generate_latex_label(r"$a$bc$_∭X∀Y"),
+            _utils.generate_latex_label(r"$a$bc$_∭X∀Y"),
         )
 
     def test_array_to_latex(self):
